Get and install NSIS unicode plugins
Uses the included Python script to download NSIS unicode plugins. Then
installs relevant components in the `%PREFIX%` for inclusion in the
package. These seem to the be the same bits that Anaconda includes.
Unfortunately `elevate.exe` has not been found yet, so is not being
included. Everything else is though.