Concatenate NSIS and plugin license files
Combines the NSIS license file and all plugin license files found into
one big license file to include in the conda package.

Could not find a license file for `UnicodePathTest`, so have left that
one out. It suggests in the README that it is under the LGPL though no
version is specified. Not much we can do about it.